ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
pydantic==2.5.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
requests==2.31.0
psycopg2-binary==2.9.9

